refactor: task-8_redux, add separate Header container
